# Optimize VM performance with pre-compiled jump table