@Wi1l-B0t 2025-06-26
Where ar  pre-compiled jump tables saved? Or need to compile every time?